pyproject.toml
[tool.poetry]
name = "password-manager"
version = "0.1.0"
description = ""
authors = ["DorLex <dorozhinskiysanya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
cryptography = "^41.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
